Move permissions to a file
